requirements.txt
alabaster==0.7.13
anyascii==0.3.2
astroid==2.15.6
Babel==2.12.1
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
docutils==0.18.1
exceptiongroup==1.2.0
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.32
h5py==3.9.0
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
mypy-extensions==1.0.0
myst-parser==2.0.0
numba==0.57.1
numpy==1.24.0
packaging==23.1
pandas==2.1.4
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.9.1
pluggy==1.3.0
py-pde==0.32.1
Pygments==2.16.1
pyparsing==3.0.9
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==7.2.5
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sympy==1.12
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
